requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
pandas>=2.2.0
scikit-learn>=1.4.0
joblib>=1.3.0
